Image Converter/requirements.txt
Flask==3.1.0
numpy==1.24.3
Pillow==9.5.0
